requirements.txt
discord.py>=1.0.0a0
